+++ backend/src/main/java/ru/focsit/backend/rest/controller/user/PlaylistRestController.java
@@ -1,12 +1,16 @@
 package ru.focsit.backend.rest.controller.user;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import ru.focsit.backend.pojo.Playlist;
 import ru.focsit.backend.pojo.Track;
+import ru.focsit.backend.pojo.User;
 import ru.focsit.backend.service.PlaylistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import ru.focsit.backend.service.TrackService;
+import ru.focsit.backend.service.UserService;
 
 import java.util.List;
 import java.util.Optional;
@@ -20,6 +24,9 @@ public class PlaylistRestController {
 
     @Autowired
     private TrackService trackService;
+
+    @Autowired
+    private UserService userService;
 
     @GetMapping("/{id}")
     public ResponseEntity<Playlist> getPlaylistById(@PathVariable Long id) {
@@ -44,14 +51,41 @@ public class PlaylistRestController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Playlist> updatePlaylist(@PathVariable Long id, @RequestBody Playlist playlistDetails) {
-        Playlist updatedPlaylist = playlistService.updatePlaylist(id, playlistDetails);
-        return updatedPlaylist != null ? ResponseEntity.ok(updatedPlaylist) : ResponseEntity.notFound().build();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String curUserName = authentication.getName();
+        User curUser = userService.findByUserLogin(curUserName);
+        Optional<Playlist> playlistOptional = playlistService.getPlaylistById(id);
+        if (playlistOptional.isPresent()) {
+            Playlist existingPlaylist = playlistOptional.get();
+            if (existingPlaylist.getPlaylistUser().equals(curUser)) {
+                playlistDetails.setPlaylistUser(curUser);
+                Playlist updatedPlaylist = playlistService.updatePlaylist(id, playlistDetails);
+                return updatedPlaylist != null ? ResponseEntity.ok(updatedPlaylist) : ResponseEntity.notFound().build();
+            } else {
+                return ResponseEntity.status(403).build(); // Forbidden
+            }
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePlaylist(@PathVariable Long id) {
-        playlistService.deletePlaylist(id);
-        return ResponseEntity.noContent().build();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String curUserName = authentication.getName();
+        User curUser = userService.findByUserLogin(curUserName);
+        Optional<Playlist> playlistOptional = playlistService.getPlaylistById(id);
+        if (playlistOptional.isPresent()) {
+            Playlist existingPlaylist = playlistOptional.get();
+            if (existingPlaylist.getPlaylistUser().equals(curUser)) {
+                playlistService.deletePlaylist(id);
+                return ResponseEntity.noContent().build();
+            } else {
+                return ResponseEntity.status(403).build(); // Forbidden
+            }
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @GetMapping("/{id}/tracks/search")
